Clean l2sql main function
This refactors the l2sql module in various ways:

a) Returns query from l2sql get query methods as we want to test
the output of the class, not the output to stdout
b) main() should only be responsible for argument parsing and we want to test the main class (l2sql) and not main()
c) Tests are now testing l2sql instead of main.
d) Removed stdout capturing logic in tests which is no longer required as it is returned
e) Moved argument parsing to use argparse instead of implementing manually.