Use Consistent Package structure and follow pep8 naming
This refactor is to convert the project to a more standard
Python Project.

Links:
* http://docs.python-guide.org/en/latest/writing/structure/
* http://as.ynchrono.us/2007/12/filesystem-structure-of-python-project_21.html

This commit:

a) moves all python files and data files to a folder with
the same name as project.
b) Uses consistent module naming styles according to pep8
(https://www.python.org/dev/peps/pep-0008/)
c) Uses relative paths in modules.
d) Fixes hardcoding in tests of database and other folders. This will be converted to use system package later to make it work on other operating systems.